Major Change to AddVar/AddVarVec + Eq() Definition and Testing for Some Vector Expressions
I modified the function AddVar so that it no longer needs to be called with lower and upper bounds known a priori (the default is gurobi.go's negative and positive infity values respectively). I also added and tested a lot of files for the Eq() functions for KVector and VarVector.